% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binw.R
\name{binw}
\alias{binw}
\title{Calculate the Bin Width for Plotting Histograms}
\usage{
binw(x, method = c("fd", "scott", "sturges"))
}
\arguments{
\item{x}{A numeric vector of the values from which histogram has to be
generated.}

\item{method}{The method to compute the number of classes for the histogram.}
}
\value{
The bin width.
}
\description{
Calculate the Bin Width for Plotting Histograms
}
\examples{
set.seed(1)
x <- stats::rnorm(1111)

binw(x = x, method = "fd")
binw(x = x, method = "scott")
binw(x = x, method = "sturges")

}
\seealso{
\code{\link[grDevices]{nclass}}
}
