% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.calc.R
\name{diversity.calc}
\alias{diversity.calc}
\title{Compute Diversity Measures}
\usage{
diversity.calc(x, base = exp(1), na.omit = TRUE)
}
\arguments{
\item{x}{A factor vector of categories (e.g., species, traits). The frequency
of each level is treated as the abundance of that category.}

\item{base}{The logarithm base to be used for computation of different
diversity indices. Default is \code{exp(1)}.}

\item{na.omit}{logical. If \code{TRUE}, missing values (\code{NA}) are
ignored and not included as a distinct factor level for computation.
Default is \code{TRUE}.}
}
\value{
A list of different \describe{ \item{richness}{The number of classes
  in \code{x} or the richness.} \item{margalef_index}{Margalef's Richness
  Index (\mjseqn{D_{Margalef}})
  \insertCite{margalef_information_1973}{avial}}
  \item{menhinick_index}{Menhinick's Index (\mjseqn{D_{Menhinick}})
  \insertCite{menhinick_comparison_1964}{avial}} \item{berger_parker}{Berger–Parker Index
  (\mjseqn{D_{BP}}) \insertCite{berger_diversity_1970}{avial}}
  \item{berger_parker_reciprocal}{Reciprocal Berger–Parker Index
  (\mjseqn{D_{BP_{R}}}) \insertCite{magurran_measuring_2011}{avial}}
  \item{simpson}{Simpson's index (\mjseqn{d})
  \insertCite{simpson_measurement_1949,peet_measurement_1974}{avial}}
  \item{gini_simpson}{Simpson's Index of Diversity or Gini's Diversity Index
  or Gini-Simpson Index or Nei's Diversity Index or Nei's Variation Index
  (\mjseqn{D}) or Hurlbert’s probability of interspecific encounter
  (\mjseqn{PIE})
  \insertCite{gini_variabilita_1912,gini_variabilita_1912-2,greenberg_measurement_1956,berger_diversity_1970,hurlbert_nonconcept_1971,nei_analysis_1973,peet_measurement_1974}{avial}}
  \item{simpson_max}{Maximum Simpson's Index of Diversity or Maximum Nei's
  Diversity/Variation Index (\mjseqn{D_{max}})
  \insertCite{hennink_interpretation_1990}{avial}}
  \item{simpson_reciprocal}{Simpson's Reciprocal Index or Hill's
  \mjseqn{N_{2}} (\mjseqn{D_{R}}) or Effective number of Species
  (\mjseqn{ENS_{d}})
  \insertCite{williams_patterns_1964,hill_diversity_1973}{avial}}
  \item{simpson_relative}{Relative Simpson's Index of Diversity or Relative
  Nei's Diversity/Variation Index (\mjseqn{D'})
  \insertCite{hennink_interpretation_1990}{avial}}
  \item{simpson_evenness}{Simpson’s evenness or equitability
  (\mjseqn{D_{e}})
  \insertCite{pielou_measurement_1966,hill_diversity_1973}{avial}}
  \item{shannon}{Shannon or Shannon-Weaver or Shannon-Wiener Diversity Index
  or Shannon entropy (\mjseqn{H})
  \insertCite{shannon_mathematical_1949,peet_measurement_1974}{avial}}
  \item{shannon_max}{Maximum Shannon-Weaver Diversity Index
  (\mjseqn{H_{max}}) \insertCite{hennink_interpretation_1990}{avial}}
  \item{shannon_relative}{Relative Shannon-Weaver Diversity Index or Shannon
  Equitability Index (\mjseqn{H'}) or Pielou's Evenness (\mjseqn{J})
  \insertCite{pielou_measurement_1966,hennink_interpretation_1990}{avial}}
  \item{shannon_ens}{Effective number of species for the Shannon - Weaver
  Diversity Index (\mjseqn{ENS_{H}}) or Hill's \mjseqn{N_{1}}
  \insertCite{macarthur_patterns_1965,hill_diversity_1973}{avial}}
  \item{heip_evenness}{Heip's Evenness Index (\mjseqn{E_{Heip}})
  \insertCite{heip_new_1974}{avial}} \item{mcintosh_diversity}{McIntosh
  Diversity Index (\mjseqn{D_{Mc}})
  \insertCite{mcintosh_index_1967,peet_measurement_1974}{avial}}
  \item{mcintosh_evenness}{McIntosh Evenness Index (\mjseqn{E_{Mc}})
  \insertCite{pielou_ecological_1975}{avial}}
  \item{smith_wilson}{Smith & Wilson's Evenness Index (\mjseqn{E_{var}})
  \insertCite{smith_consumers_1996}{avial}}
  \item{brillouin_index}{Brillouin Diversity Index
  \insertCite{brillouin_science_2013}{avial}}
  \item{renyi_entropy_0}{Rényi Entropy of order 0
  \insertCite{renyi_measures_1960}{avial}}
  \item{renyi_entropy_1}{Rényi Entropy of order 1}
  \item{renyi_entropy_2}{Rényi Entropy of order 2}
  \item{tsallis_entropy_0}{Tsallis Entropy of order 0
  \insertCite{havrda_quantification_1967,daroczy_generalized_1970,tsallis_possible_1988}{avial}}
  \item{tsallis_entropy_1}{Tsallis Entropy of order 1}
  \item{tsallis_entropy_2}{Tsallis Entropy of order 2}
  \item{hill_number_0}{Hill Number of order 0 (\mjseqn{{}^0 D})
  \insertCite{hill_diversity_1973}{avial}} \item{hill_number_1}{Hill Number
  of order 1 (\mjseqn{{}^1 D})} \item{hill_number_2}{Hill Number of order 2
  (\mjseqn{{}^2 D})}}
}
\description{
The function \code{diversity.calc} calculates the following diversity
measures \itemize{ \item{Margalef's Richness Index (\mjseqn{D_{Margalef}})
\insertCite{margalef_information_1973}{avial}} \item{Menhinick's Index
(\mjseqn{D_{Menhinick}}) \insertCite{menhinick_comparison_1964}{avial}}
\item{Berger–Parker Index (\mjseqn{D_{BP}})
\insertCite{berger_diversity_1970}{avial}} \item{Reciprocal Berger–Parker
Index (\mjseqn{D_{BP_{R}}}) \insertCite{magurran_measuring_2011}{avial}}
\item{Simpson's Index (\mjseqn{d})
\insertCite{simpson_measurement_1949,peet_measurement_1974}{avial}}
\item{Simpson's Index of Diversity or Gini's Diversity Index or Gini-Simpson
Index or Nei's Diversity Index or Nei's Variation Index (\mjseqn{D}) or
Hurlbert’s probability of interspecific encounter (\mjseqn{PIE})
\insertCite{gini_variabilita_1912,gini_variabilita_1912-2,greenberg_measurement_1956,berger_diversity_1970,hurlbert_nonconcept_1971,nei_analysis_1973,peet_measurement_1974}{avial}}
\item{Maximum Simpson's Index of Diversity or Maximum Nei's
Diversity/Variation Index (\mjseqn{D_{max}})
\insertCite{hennink_interpretation_1990}{avial}} \item{Simpson's
Reciprocal Index or Hill's \mjseqn{N_{2}} (\mjseqn{D_{R}}) or Effective
number of Species (\mjseqn{ENS_{d}})
\insertCite{williams_patterns_1964,hill_diversity_1973}{avial}}
\item{Relative Simpson's Index of Diversity or Relative Nei's
Diversity/Variation Index (\mjseqn{D'})
\insertCite{hennink_interpretation_1990}{avial}} \item{Simpson’s
evenness or equitability (\mjseqn{D_{e}})
\insertCite{pielou_measurement_1966,hill_diversity_1973}{avial}}
\item{Shannon or Shannon-Weaver or Shannon-Wiener Diversity Index or Shannon entropy
(\mjseqn{H}) \insertCite{shannon_mathematical_1949,peet_measurement_1974}{avial}}
\item{Maximum Shannon-Weaver Diversity Index (\mjseqn{H_{max}})
\insertCite{hennink_interpretation_1990}{avial}} \item{Relative
Shannon-Weaver Diversity Index or Shannon Equitability Index (\mjseqn{H'})
or Pielou's Evenness (\mjseqn{J})
\insertCite{pielou_measurement_1966,hennink_interpretation_1990}{avial}}
\item{Effective number of species for the Shannon - Weaver Diversity Index
(\mjseqn{ENS_{H}}) or Hill's \mjseqn{N_{1}}
\insertCite{macarthur_patterns_1965,hill_diversity_1973}{avial}} \item{Heip's
Evenness Index (\mjseqn{E_{Heip}}) \insertCite{heip_new_1974}{avial}}
\item{McIntosh Diversity Index (\mjseqn{D_{Mc}})
\insertCite{mcintosh_index_1967,peet_measurement_1974}{avial}}
\item{McIntosh Evenness Index (\mjseqn{E_{Mc}})
\insertCite{pielou_ecological_1975}{avial}} \item{Smith & Wilson's Evenness
Index (\mjseqn{E_{var}}) \insertCite{smith_consumers_1996}{avial}}
\item{Brillouin Diversity Index (\mjseqn{D_{Brillouin}})
\insertCite{brillouin_science_2013}{avial}} \item{Rényi Entropy
(\mjseqn{{}^q H_{Rényi}}) \insertCite{renyi_measures_1960}{avial}}
\item{Tsallis or HCDT Entropy (\mjseqn{{}^q H_{Tsallis}})
\insertCite{havrda_quantification_1967,daroczy_generalized_1970,tsallis_possible_1988}{avial}}
\item{Hill Numbers (\mjseqn{{}^q D})
\insertCite{hill_diversity_1973}{avial}} } \loadmathjax
}
\section{Details}{
 The diversity indices implemented in \code{diversity.calc}
  are computed as follows.

  \subsection{Richness Indices}{

  The number of classes of a phenotypic trait (or species richness)
  (\mjseqn{k}) can be described by adjusting for sample size (\mjseqn{N}) in
  Margalef's Richness Index (\mjseqn{D_{Margalef}})
  \insertCite{margalef_information_1973}{avial} and Menhinick's Index
  (\mjseqn{D_{Menhinick}}) \insertCite{menhinick_comparison_1964}{avial}

  These are computed as follows.

   \mjsdeqn{D_{Margalef} = \frac{k - 1}{\ln(N)}}

    \mjsdeqn{D_{Menhinick} = \frac{k}{\sqrt{N}}}

  }

  \subsection{Berger–Parker Index}{ This is the is the proportion of
  individuals belonging to the most abundant class in a trait (or species in
  a community) and is computed as below.

  \mjsdeqn{D_{BP} = \max(p_i)}

  Where, \mjseqn{p_{i}} denotes the proportion/fraction/frequency of
  accessions in the \mjseqn{i}th phenotypic class for a trait (or number
  of individuals in the \mjseqn{i}th species).

  It's reciprocal estimates the relative diversity of this class.

  \mjsdeqn{D_{BP_{R}} = \frac{1}{D_{BP}}}

  }

  \subsection{Simpson's and Related Indices}{Simpson's index (\mjseqn{d})
  which estimates the probability that two accessions randomly selected will
  belong to the same phenotypic class of a trait (or species in a
  community), is computed as follows
  \insertCite{simpson_measurement_1949,peet_measurement_1974}{avial}.

  \mjsdeqn{d = \sum_{i = 1}^{k}p_{i}^{2}}

  Where, \mjseqn{k} is the number of phenotypic classes for the trait (or
  number of species in the community).

  The value of \mjseqn{d} can range from 0 to 1 with 0 representing maximum
  diversity and 1, no diversity.

  \mjseqn{d} is subtracted from 1 to give Simpson's index of diversity
  (\mjseqn{D})
  \insertCite{greenberg_measurement_1956,berger_diversity_1970,hurlbert_nonconcept_1971,peet_measurement_1974,hennink_interpretation_1990}{avial}
   originally suggested by
  \insertCite{gini_variabilita_1912,gini_variabilita_1912-2;textual}{avial}
   and described in literature as Gini's diversity index or Gini-Simpson
  index. It is the same as Nei's diversity index or Nei's variation index
  \insertCite{nei_analysis_1973,hennink_interpretation_1990}{avial}.
  Greater the value of \mjseqn{D}, greater the diversity with a range from 0
  to 1.

  \mjsdeqn{D = 1 - d}

  The maximum value of \mjseqn{D}, \mjseqn{D_{max}} occurs when accessions
  are uniformly distributed across the phenotypic classes (or individuals are
  uniformly distributed across species in a community) and is computed as
  follows \insertCite{hennink_interpretation_1990}{avial}.

  \mjsdeqn{D_{max} = 1 - \frac{1}{k}}

  Reciprocal of \mjseqn{d} gives the Simpson's reciprocal index
  (\mjseqn{D_{R}})
  \insertCite{williams_patterns_1964,hennink_interpretation_1990}{avial}
   and can range from 1 to \mjseqn{k}. This was also described in
  \insertCite{hill_diversity_1973;textual}{avial} as \mjseqn{N_{2}} or
  as Effective number of Species (or classes) (\mjseqn{ENS_{d}}).

  \mjsdeqn{D_{R} = \frac{1}{d}}

  Relative Simpson's index of diversity or Relative Nei's diversity/variation
  index (\mjseqn{H'}) \insertCite{hennink_interpretation_1990}{avial}
  is defined as follows \insertCite{peet_measurement_1974}{avial}.

  \mjsdeqn{D' = \frac{D}{D_{max}}}

  Simpson’s evenness or equitability (\mjseqn{D_{e}} is described as follows
  \insertCite{pielou_measurement_1966,hill_diversity_1973}{avial}.

  \mjsdeqn{D_{e} = \frac{1}{d \cdot k}}

  }

  \subsection{Shannon-Weaver and Related Indices}{An index of information
  \mjseqn{H}, was described by
  \insertCite{shannon_mathematical_1949;textual}{avial} as follows.

  \mjsdeqn{H = -\sum_{i=1}^{k}p_{i} \log_{2}(p_{i})}

  \mjseqn{H} is described as Shannon or Shannon-Weaver or Shannon-Wiener
  diversity index or Shannon entropy in literature
  \insertCite{shannon_mathematical_1949,peet_measurement_1974}{avial}.

  Alternatively, \mjseqn{H} is also computed using natural logarithm instead
  of logarithm to base 2.

  \mjsdeqn{H = -\sum_{i=1}^{k}p_{i} \ln(p_{i})}

  The maximum value of \mjseqn{H} (\mjseqn{H_{max}}) is \mjseqn{\ln(k)}. This
  value occurs when each phenotypic class for a trait has the same proportion
  of accessions (or each species in a community has the same proportion of
  individuals) \insertCite{hennink_interpretation_1990}{avial}.

  \mjsdeqn{H_{max} = \log_{2}(k)\;\; \textrm{OR} \;\; H_{max} = \ln(k)}

  The relative Shannon-Weaver diversity index or Shannon equitability index
  (\mjseqn{H'}) or Pielou's Evenness (\mjseqn{J}) is the Shannon diversity
  index (\mjseqn{I}) divided by the maximum diversity (\mjseqn{H_{max}})
  \insertCite{pielou_measurement_1966,hennink_interpretation_1990}{avial}.

  \mjsdeqn{H' = \frac{H}{H_{max}}}

  \insertCite{macarthur_patterns_1965;textual}{avial} described the Effective
  number of species (or classes) for the Shannon index (\mjseqn{ENS_{H}}) as
  follows.

  \mjsdeqn{ENS_{H} = e^{H}}

  Heip’s index or Heip's Evenness index is a transformation of the
  Shannon–Wiener diversity index that standardizes it relative to number of
  classes in the trait (or species richness) and is computed as follows
  \insertCite{heip_new_1974}{avial}.

  \mjsdeqn{E_{Heip} = \frac{e^{H} - 1}{k - 1}}

  }

  \subsection{McIntosh's Measure of Diversity}{A similar index of diversity
  was described by \insertCite{mcintosh_index_1967;textual}{avial} as
  follows (\mjseqn{D_{Mc}}) \insertCite{peet_measurement_1974}{avial}.

  \mjsdeqn{D_{Mc} = \frac{N - \sqrt{\sum_{i=1}^{k}n_{i}^2}}{N - \sqrt{N}}}

  Where, \mjseqn{n_{i}} denotes the number of accessions in the \mjseqn{i}th
  phenotypic class for a trait (or number of individuals in the \mjseqn{i}th
  species in the community) and \mjseqn{N} is the total number of accessions
  so that \mjseqn{p_{i} = {n_{i}}/{N}}.

  An additional measure of evenness was proposed by
  \insertCite{pielou_ecological_1975;textual}{avial} as follows.

  \mjsdeqn{E_{Mc} = \frac{N - \sqrt{\sum_{i=1}^{k}n_{i}^2}}{N -
  \frac{N}{\sqrt{S}}}}

  }

  \subsection{Smith & Wilson's Evenness Index}{This index measures how
  equally are accessions/genotypes distributed among different trait classes
  (or individuals individuals are distributed among different species in a
  community). This is less sensitive to rare classes or species and is
  computed as follows.

  \mjsdeqn{E_{var} = 1 - \frac{2}{\pi} \arctan{\left ( \frac{1}{k}
  \sum_{i=1}^{k}(\ln{n_{i} - \overline{\ln{n}}})^{2} \right )}}

  }

  \subsection{Brillouin Diversity Index}{This is an information-theoretic
  measure appropriate for complete censuses and is computed as follows
  \insertCite{brillouin_science_2013}{avial}.

  \mjsdeqn{H_{B} = \frac{\ln(N!) - \sum \ln(n_i!)}{N}}

  }

  \subsection{Parametric Indices}{

  Parametric indices, also known as multivariate or compound indices, use a
  sensitivity parameter (\mjseqn{q}) to weigh frequent and rare classes
  within a trait (or common or rare species within a community).

  The Rényi entropy extends several entropy measures, including Shannon
  entropy, and is computed as follows
  \insertCite{renyi_measures_1960}{avial}.

  \mjsdeqn{{}^q H_{Rényi} = \frac{1}{1-q} \ln \sum_{i=1}^{k} p_{i}^{q} ,
  \quad q \ge 0, q \neq 1}

  It is more frequently computed using natural logarithm instead of
  logarithm to base 2. The index is undefined for (\mjseqn{q = 1}), but
  Shannon entropy is as a limiting case.

  Tsallis proposed a similar measure, the HCDT or Tsallis entropy
  \insertCite{havrda_quantification_1967,daroczy_generalized_1970,tsallis_possible_1988}{avial},
  which matches species richness for \mjseqn{q = 0}, Shannon entropy
  \mjseqn{q = 1}, and the Gini-Simpson index for \mjseqn{q = 2}.

  \mjsdeqn{{}^q H_{Tsallis} = \frac{1}{q - 1} \left (
  1 - \sum_{i=1}^{k} p_i^q \right ), \quad q \ge 0, q \neq 1}

  Hill showed that species richness, Shannon entropy, and Simpson's
  index are all related diversity indices, collectively known
  as Hill numbers which is defined as below
  \insertCite{hill_diversity_1973}{avial}.

  \mjsdeqn{{}^q D = {\left ( \sum_{i=1}^{k} p_{i}^{q}
  \right )}^{\frac{1}{1-q}} , \quad q \ge 0, q \neq 1}

 Where,

  \mjsdeqn{{}^0 D = k}

  \mjsdeqn{{}^1 D = e^{H}}

  \mjsdeqn{{}^2 D = D_{R}}

  }
}

\examples{
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Qualitative trait data ----
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

library(EvaluateCore)

pdata <- cassava_CC

qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

# Convert qualitative data columns to factor
pdata[, qual] <- lapply(pdata[, qual], as.factor)

# Get diversity measures
diversity.calc(x = pdata$CUAL)

# Get diversity measures for multiple qualitative traits
div_out1 <- lapply(pdata[, qual], diversity.calc)
do.call(rbind, div_out1)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Species abundance data ----
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

library(vegan)
data(dune)

abundance_site1 <- unlist(dune[1, ])
abundance_site1[abundance_site1 != 0]

# Convert to raw counts for use in diversity.calc using rep
abundance_site1_raw <- factor(rep(names(abundance_site1),
                                  times = abundance_site1))

# Get diversity measures
diversity.calc(x = abundance_site1_raw)

# Convert multiple site abundance data to raw counts
abundance_site_raw <- apply(dune, 1, function(x) {
  factor(rep(names(x), times = x))
})

# Get diversity measures for multiple sites
div_out2 <- lapply(abundance_site_raw, diversity.calc)
do.call(rbind, div_out2)

}
\references{
\insertAllCited{}
}
