% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.compare.R
\name{diversity.compare}
\alias{diversity.compare}
\title{Compare Diversity Measures}
\usage{
diversity.compare(
  x,
  group,
  R = 1000,
  base = exp(1),
  na.omit = TRUE,
  p.adjust.method = c("bonferroni", "holm"),
  ci.conf = 0.95,
  ci.type = c("perc", "bca"),
  q = seq(0, 3, 0.1),
  parallel = c("no", "multicore", "snow"),
  ncpus = 1L,
  cl = NULL
)
}
\arguments{
\item{x}{A factor vector of categories (e.g., species, traits). The frequency
of each level is treated as the abundance of that category.}

\item{group}{A factor vector indicating the group of each observation. Must
have the same length as \code{x}.}

\item{R}{Integer specifying the number of permutations. Default is 1000.}

\item{base}{The logarithm base to be used for computation of shannon family
of diversity indices. Default is \code{exp(1)}.}

\item{na.omit}{logical. If \code{TRUE}, missing values (\code{NA}) are
ignored and not included as a distinct factor level for computation.
Default is \code{TRUE}.}

\item{p.adjust.method}{(perm.test.pairwise only) Method for adjusting
p-values for multiple comparisons. Options include \code{"bonferroni"} and
\code{"holm"}. Default is \code{"bonferroni"}.}

\item{ci.conf}{Confidence level of the bootstrap interval. Default is 0.95.}

\item{ci.type}{A vector of character strings representing the type of
intervals required. The options are \code{c("perc", "bca")}.}

\item{q}{The order of the parametric index.}

\item{parallel}{
    The type of parallel operation to be used (if any).  If missing, the
    default is taken from the option \code{"boot.parallel"} (and if that
    is not set, \code{"no"}).
  }

\item{ncpus}{
    integer: number of processes to be used in parallel operation:
    typically one would chose this to the number of available CPUs.
  }

\item{cl}{
    An optional \pkg{parallel} or \pkg{snow} cluster for use if
    \code{parallel = "snow"}.  If not supplied, a cluster on the
    local machine is created for the duration of the \code{boot} call.
  }
}
\description{
Compare Diversity Measures
}
\examples{
library(EvaluateCore)

pdata <- cassava_CC

qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

# Convert qualitative data columns to factor
pdata[, qual] <- lapply(pdata[, qual], as.factor)

str(pdata)

diversity.compare(x = pdata$CUAL, group = pdata$LNGS, R = 100,
                  base = exp(1), na.omit = TRUE)

diversity.compare(x = pdata$ANGB, group = pdata$LNGS, R = 100,
                  base = exp(1), na.omit = TRUE)

}
