% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.profile.R
\name{diversity.profile}
\alias{diversity.profile}
\title{Generate Diversity Profiles for Parametric Indices}
\usage{
diversity.profile(
  x,
  group,
  q = seq(0, 3, 0.1),
  conf = 0.95,
  R = 1000,
  parameter = c("hill", "renyi", "tsallis"),
  ci.type = c("perc", "bca"),
  parallel = c("no", "multicore", "snow"),
  ncpus = getOption("boot.ncpus", 1L),
  cl = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric or factor vector of observations.}

\item{group}{A factor vector indicating the group of each observation. Must
have the same length as \code{x}.}

\item{q}{The order of the parametric index.}

\item{conf}{Confidence level of the interval. Default is 0.95.}

\item{R}{Integer specifying the number of permutations. Default is 1000.}

\item{parameter}{The parametric index. Options include \code{"hill"},
\code{"renyi"} and \code{"tsallis"}. Default is \code{"hill"}.}

\item{ci.type}{A vector of character strings representing the type of
intervals required. The value should be any subset of the values
\code{c("norm", "basic", "stud", "perc", "bca")} or simply \code{"all"}
which will compute all five types of intervals.}

\item{parallel}{
    The type of parallel operation to be used (if any).  If missing, the
    default is taken from the option \code{"boot.parallel"} (and if that
    is not set, \code{"no"}).
  }

\item{ncpus}{
    integer: number of processes to be used in parallel operation:
    typically one would chose this to the number of available CPUs.
  }

\item{cl}{
    An optional \pkg{parallel} or \pkg{snow} cluster for use if
    \code{parallel = "snow"}.  If not supplied, a cluster on the
    local machine is created for the duration of the \code{boot} call.
  }

\item{...}{}
}
\value{
A list of data frames with the following columns for each factor
  level in \code{group}. \describe{ \item{q}{} \item{observed}{}
  \item{mean}{} \item{lower}{} \item{upper}{} }
}
\description{
Generate Diversity Profiles for Parametric Indices
}
