% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_functions.R
\name{diversity_functions}
\alias{diversity_functions}
\alias{berger_parker}
\alias{berger_parker_reciprocal}
\alias{simpson}
\alias{gini_simpson}
\alias{simpson_max}
\alias{simpson_reciprocal}
\alias{simpson_relative}
\alias{simpson_evenness}
\alias{shannon}
\alias{shannon_max}
\alias{shannon_relative}
\alias{shannon_ens}
\alias{mcintosh_diversity}
\alias{mcintosh_evenness}
\alias{smith_wilson}
\alias{heip_evenness}
\alias{margalef_index}
\alias{menhinick_index}
\alias{brillouin_index}
\alias{hill_number}
\alias{renyi_entropy}
\alias{tsallis_entropy}
\alias{hill_evenness}
\title{Diversity Index Functions}
\usage{
berger_parker(x)

berger_parker_reciprocal(x)

simpson(x)

gini_simpson(x)

simpson_max(x)

simpson_reciprocal(x)

simpson_relative(x)

simpson_evenness(x)

shannon(x, base = 2)

shannon_max(x, base = 2)

shannon_relative(x, base = 2)

shannon_ens(x, base = 2)

mcintosh_diversity(x)

mcintosh_evenness(x)

smith_wilson(x)

heip_evenness(x)

margalef_index(x)

menhinick_index(x)

brillouin_index(x)

hill_number(x, q = 1)

renyi_entropy(x, q = 1)

tsallis_entropy(x, q = 1)

hill_evenness(x, q = 1)
}
\arguments{
\item{x}{A factor vector of categories (e.g., species, traits). The frequency
of each level is treated as the abundance of that category.}

\item{base}{The logarithm base to be used for computation of shannon family
of diversity indices. Default is \code{exp(1)}.}
}
\value{
The calculated diversity index value.
}
\description{
To be used by \code{\link[avial]{diversity.calc}}.
}
