% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_ggplot.R
\name{dnorm_ggplot}
\alias{dnorm_ggplot}
\title{Function to Add Normal Distribution Curve in \code{\link[ggplot2]{ggplot2}}
histogram}
\usage{
dnorm_ggplot(x, mean, sd, n, bw)
}
\arguments{
\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{n}{The number of records or data points used to plot the histogram.}

\item{bw}{The bin width of the histogram.}
}
\value{
The density for normal distribution.
}
\description{
Enhancement of \code{\link[stats]{dnorm}} to plot normal distribution or bell
curve as an overlay in \code{\link[ggplot2]{ggplot2}} histograms according to
number of records and bin width.
}
\examples{
dnorm(0) * 25 == dnorm_ggplot(0, mean = 0, sd = 1, n = 5, bw = 5)
dnorm(1) * 21 == dnorm_ggplot(1, mean = 0, sd = 1, n = 7, bw = 3)

}
\seealso{
\code{\link[ggplot2]{geom_histogram}},
  \code{\link[avial]{groupwise_histogram}}
}
