% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ems.bartlett.test.R
\name{ems.bartlett.test}
\alias{ems.bartlett.test}
\title{Bartlett Test of Homogeneity of Error Variances}
\usage{
ems.bartlett.test(ems, df)
}
\arguments{
\item{ems}{A vector of error variances or error mean sum of squares from each
environment (years and/or locations).}

\item{df}{A vector of degrees of freedom correspondint to \code{ems}.}
}
\value{
A list with the chi-square value test statistic, corresponding
  degrees of freedom and p value.
}
\description{
Bartlett Test of Homogeneity of Error Variances
}
\examples{
# Example from Page 481 Gomez KA and AA Gomez (1984) Statistical Procedures
# for Agricultural Research. 2nd ed. Wiley, New York, 680 p.

# Different degrees of freedom
ems <- c(6.73920, 1.93496, 1.15500, 10.58450)
df <- c(19, 16, 17, 19)

ems.bartlett.test(ems = c(6.73920, 1.93496, 1.15500, 10.58450),
                  df = c(19, 16, 17, 19))

# Same degrees of freedom
ems <- c(11.459848, 17.696970, 10.106818)
df <- c(20, 20, 20)

ems.bartlett.test(ems = c(11.459848, 17.696970, 10.106818),
                  df = c(20, 20, 20))
}
\seealso{
\code{\link[stats]{bartlett.test}}
}
