% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_constrained_order.R
\name{group_constrained_order}
\alias{group_constrained_order}
\title{Ordering by keeping defined groups together}
\usage{
group_constrained_order(x, groups, group_order = c("groups", "original"))
}
\arguments{
\item{x}{The original vector of elements in order.}

\item{groups}{The groups of elements that are to be kept together as a list
of vectors. The groups must be disjoint.}

\item{group_order}{The group internal order should follow the one in
\code{"groups"} or the \code{"original"} order.}
}
\value{
A vector of the original elements where the specified groups of
  elements are kept together.
}
\description{
This function modifies the order of a vector so that specified groups of
elements remain together. This is useful for keeping specific dendrogram
leaves together. The location of each group block is determined by the
earliest member of that group in original vector. For elements not involved
in any group, their relative order is unchanged.
}
\examples{
set.seed(123)

# data
mat <- matrix(rnorm(8 * 4), nrow = 8)
rownames(mat) <- LETTERS[1:8]

# hierarchical clustering
hc <- hclust(dist(mat))
plot(hc, main = "Original dendrogram")

# groups of leaves to keep together (may not be adjacent in dendrogram)
groups <- list(
  g1 = c("B", "F"),
  g2 = c("C", "D", "E")
)

# Dendrogram original order
orig_order <- hc$labels[hc$order]
orig_order

# Get group constrained order
new_order <-
  group_constrained_order(x = orig_order,
                          groups = groups,
                          group_order = "original")

new_order

# Reorder dendrogram labels
hc2 <- hc
hc2$order <- match(new_order, hc$labels)

plot(hc2, main = "Dendrogram with grouped leaves")

}
