% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwise_dumbell.R
\name{groupwise_dumbell}
\alias{groupwise_dumbell}
\title{Plot Group-wise Dumbell Plots}
\usage{
groupwise_dumbell(
  data,
  group,
  trait,
  genotype,
  subset = c("facet", "list", "none"),
  diff.sort = c("none", "ascending", "descending"),
  segment = TRUE,
  segment.size = 3,
  segment.colour = "gray",
  segment.alpha = 0.5,
  point.size = 3,
  point.alpha = 0.8,
  error.bar = TRUE,
  error.bar.width = 0.1,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
columns specifying the group and trait.}

\item{group}{Name of column specifying the group as a character string.}

\item{trait}{Name of column specifying the trait as a character string.}

\item{genotype}{Name of column specifying the genotype as a character string.}

\item{subset}{The method for subsetting the plots according to the argument
\code{"group"}. Either \code{"facet"} for getting an plot using faceting in
\code{ggplot2} or \code{"list"} for getting a list of plots.}

\item{diff.sort}{The order for sorting the genotypes for plotting. Either
\code{"ascending"}, \code{"descending"} or \code{"none"}.}

\item{segment}{logical. If \code{TRUE}, the dumbell segment is plotted.
Default is \code{TRUE}.}

\item{segment.size}{The size of the dumbell segment.}

\item{segment.colour}{The colour of the dumbell segment.}

\item{segment.alpha}{Alpha transparency for the dumbell segment.}

\item{point.size}{The size of the points.}

\item{point.alpha}{Alpha transparency for the points.}

\item{error.bar}{logical. If \code{TRUE}, the error bars depicting standard
errors are plotted. Default is \code{TRUE}.}

\item{error.bar.width}{The width of the error bars.}

\item{ncol}{Number of columns when \code{subset = "facet"}.}

\item{nrow}{Number of rows when \code{subset = "facet"}.}
}
\value{
The group-wise dumbell plot as a \code{ggplot2} plot grob or as a
  list of \code{ggplot2} plot grobs.
}
\description{
Plot Group-wise Dumbell Plots
}
\examples{
library(agridat)
library(ggplot2)
library(patchwork)

soydata <- australia.soybean

checks <- c("G01", "G05")

checkdata <- soydata[soydata$gen \%in\% checks, ]

clrs <- c("#B2182B", "#2166AC", "#009E53", "#E69F00")
clrs_dark <- colorspace::darken(clrs, amount = 0.2)

# Group-wise dumbell plot with error bar
outg_dumbell1 <-
  groupwise_dumbell(data = checkdata, group = "loc",
                    trait = "lodging", genotype = "gen",
                    subset = "none", diff.sort = "descending")

outg_dumbell1 +
  scale_colour_manual(values = clrs)

# Group-wise dumbell plot without error bar
outg_dumbell2 <-
  groupwise_dumbell(data = checkdata, group = "loc",
                    trait = "lodging", genotype = "gen",
                    subset = "none", diff.sort = "descending",
                    error.bar = FALSE)

outg_dumbell2 +
  scale_colour_manual(values = clrs)

# Group-wise points with error bar as facets
outg_facet <-
  groupwise_dumbell(data = checkdata, group = "loc",
                    trait = "lodging", genotype = "gen",
                    subset = "facet")

outg_facet +
  scale_colour_manual(values = clrs)

# Group-wise points with error bar as list
outg_list <-
  groupwise_dumbell(data = checkdata, group = "loc",
                    trait = "lodging", genotype = "gen",
                    subset = "list")

wrap_plots(outg_list, nrow = 2, guides = "collect")

outg_list <-
  lapply(seq_along(outg_list), function(i) {
    outg_list[[i]] +
      scale_colour_manual(values = clrs[i])
  })

wrap_plots(outg_list, nrow = 2, guides = "collect")

}
