% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwise_histogram.R
\name{groupwise_histogram}
\alias{groupwise_histogram}
\title{Plot Group-wise Histograms}
\usage{
groupwise_histogram(
  data,
  group,
  trait,
  background.hist = TRUE,
  background.hist.alpha = 0.25,
  background.density = TRUE,
  background.density.alpha = 0.1,
  hist = TRUE,
  hist.border = TRUE,
  hist.position = c("identity", "stack"),
  hist.alpha = 0.8,
  bw.adjust = 0.5,
  density = TRUE,
  density.alpha = 0.1,
  normal.curve = TRUE,
  normal.curve.linetype = "solid",
  highlight.mean = TRUE,
  show.counts = TRUE,
  count.text.size = 3,
  subset = c("facet", "list", "none"),
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
columns specifying the group and trait.}

\item{group}{Name of column specifying the group as a character string.}

\item{trait}{Name of column specifying the trait as a character string.}

\item{background.hist}{logical. If \code{TRUE}, the background data histogram
is plotted. Default is \code{TRUE}.}

\item{background.hist.alpha}{Alpha transparency for the background data
histogram.}

\item{background.density}{logical. If \code{TRUE}, the background data kernel
density is plotted. Default is \code{TRUE}.}

\item{background.density.alpha}{Alpha transparency for the background data
kernel density.}

\item{hist}{logical. If \code{TRUE}, the group-wise histogram is plotted.
Default is \code{TRUE}.}

\item{hist.border}{logical. If \code{TRUE}, histogram border is also plotted.
Default is \code{TRUE}.}

\item{hist.position}{Histogram position adjustment. Either "identity" or
"stack".}

\item{hist.alpha}{Alpha transparency for the group-wise histogram.}

\item{bw.adjust}{Multiplicative bin width adjustment. Default is 0.5 which
means use half of the default bandwidth.}

\item{density}{logical. If \code{TRUE}, the group-wise kernel density is
plotted. Default is \code{TRUE}.}

\item{density.alpha}{Alpha transparency for the group-wise kernel density}

\item{normal.curve}{logical. If \code{TRUE}, a normal curve is plotted.
Default is \code{TRUE}.}

\item{normal.curve.linetype}{Linetype for the normal curve. See
\code{\link[ggplot2]{aes_linetype_size_shape}}.}

\item{highlight.mean}{logical. If \code{TRUE}, the mean value is highlighted
as a vertical line. Default is \code{TRUE}.}

\item{show.counts}{logical. If \code{TRUE}, group wise counts are plotted as
a text annotation. Default is \code{TRUE}.}

\item{count.text.size}{The size of the count text annotation.}

\item{subset}{The method for subsetting the plots according to the argument
\code{"group"}. Either \code{"facet"} for getting an plot using faceting in
\code{ggplot2} or \code{"list"} for getting a list of plots.}

\item{ncol}{Number of columns when \code{subset = "facet"}.}

\item{nrow}{Number of rows when \code{subset = "facet"}.}
}
\value{
The group-wise histogram as a \code{ggplot2} plot grob or as a list
  of \code{ggplot2} plot grobs.
}
\description{
Plot Group-wise Histograms
}
\examples{
library(agridat)
library(ggplot2)
library(patchwork)

soydata <- australia.soybean
# soydata[soydata$loc == "Nambour", ]$lodging <- NA

clrs <- c("#B2182B", "#2166AC", "#009E53", "#E69F00")
clrs_dark <- colorspace::darken(clrs, amount = 0.2)

# Group-wise histogram ----
outg_hist <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = FALSE,
                      background.density = FALSE,
                      hist.alpha = 0.5,
                      density = FALSE,
                      subset = "none")
outg_hist

outg_hist +
  scale_fill_manual(values = clrs) +
  scale_colour_manual(values = clrs_dark)

# Group-wise histogram - stacked ----
outg_hist_stack <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = FALSE,
                      background.density = FALSE,
                      hist.position = "stack",
                      density = FALSE,
                      normal.curve = FALSE,
                      subset = "none")
outg_hist_stack

outg_hist_stack +
  scale_fill_manual(values = clrs) +
  scale_colour_manual(values = clrs_dark)

# Group-wise histogram with facet ----
outg_hist_facet <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = TRUE,
                      background.density = FALSE,
                      hist.alpha = 0.5,
                      density = FALSE,
                      subset = "facet")
outg_hist_facet

outg_hist_facet +
  scale_fill_manual(values = clrs) +
  scale_colour_manual(values = clrs_dark)

# Group-wise histogram as list ----
outg_hist_list <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = TRUE,
                      background.density = FALSE,
                      hist.alpha = 0.5,
                      density = FALSE,
                      subset = "list")

wrap_plots(outg_hist_list, nrow = 2, guides = "collect")

outg_hist_list <-
  lapply(seq_along(outg_hist_list), function(i) {
    outg_hist_list[[i]] +
      scale_fill_manual(values = clrs[i]) +
      scale_colour_manual(values = clrs_dark[i])
  })

wrap_plots(outg_hist_list, nrow = 2, guides = "collect")

# Group-wise density ----
outg_density <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = FALSE,
                      background.density = TRUE,
                      hist = FALSE,
                      density = TRUE,
                      normal.curve = FALSE,
                      subset = "none")
outg_density

outg_density +
  scale_fill_manual(values = clrs) +
  scale_colour_manual(values = clrs_dark)

# Group-wise density with facet ----
outg_density_facet <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = FALSE,
                      background.density = TRUE,
                      hist = FALSE,
                      density = TRUE,
                      normal.curve = FALSE,
                      subset = "facet")
outg_density_facet

outg_density_facet +
  scale_fill_manual(values = clrs) +
  scale_colour_manual(values = clrs_dark)

# Group-wise density as list ----
outg_density_list <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = FALSE,
                      background.density = TRUE,
                      hist = FALSE,
                      density = TRUE,
                      normal.curve = FALSE,
                      subset = "list")

wrap_plots(outg_density_list, nrow = 2, guides = "collect")

outg_density_list <-
  lapply(seq_along(outg_density_list), function(i) {
    outg_density_list[[i]] +
      scale_fill_manual(values = clrs[i]) +
      scale_colour_manual(values = clrs_dark[i])
  })

wrap_plots(outg_density_list, nrow = 2, guides = "collect")

# Group-wise density + histogram ----
outg_density_hist <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = FALSE,
                      background.density = FALSE,
                      hist = TRUE,
                      hist.alpha = 0.3,
                      density = TRUE,
                      normal.curve = FALSE,
                      subset = "none")
outg_density_hist

outg_density_hist +
  scale_fill_manual(values = clrs) +
  scale_colour_manual(values = clrs_dark)

# Group-wise density + histogram with facet ----
outg_density_hist_facet <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = TRUE,
                      background.density = FALSE,
                      hist = TRUE,
                      hist.alpha = 0.3,
                      density = TRUE,
                      normal.curve = FALSE,
                      subset = "facet")
outg_density_hist_facet

outg_density_hist_facet +
  scale_fill_manual(values = clrs) +
  scale_colour_manual(values = clrs_dark)

# Group-wise density + histogram as list ----
outg_density_hist_list <-
  groupwise_histogram(data = soydata, group = "loc", trait = "lodging",
                      background.hist = TRUE,
                      background.density = FALSE,
                      hist = TRUE,
                      hist.alpha = 0.3,
                      density = TRUE,
                      normal.curve = FALSE,
                      subset = "list")

wrap_plots(outg_density_hist_list, nrow = 2, guides = "collect")

outg_density_hist_list <-
  lapply(seq_along(outg_density_hist_list), function(i) {
    outg_density_hist_list[[i]] +
      scale_fill_manual(values = clrs[i]) +
      scale_colour_manual(values = clrs_dark[i])
  })

wrap_plots(outg_density_hist_list, nrow = 2, guides = "collect")

}
