% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.bartlett.test.R
\name{mse.bartlett.test}
\alias{mse.bartlett.test}
\title{Bartlett's Test of Homogeneity of Error Variances}
\usage{
mse.bartlett.test(mse, df)
}
\arguments{
\item{mse}{A vector of error variances or mean squared errors from each
environment (years and/or locations).}

\item{df}{A vector of degrees of freedom corresponding to \code{mse}.}
}
\value{
A list with the chi-square value test statistic, corresponding
  degrees of freedom and p value.
}
\description{
Perform chi-square test for homogeneity of variance (Bartlett's test) to test
equality of several error variances or mean squared errors.
}
\examples{
# Examples from Page 467-471 Gomez KA and AA Gomez (1984) Statistical
# Procedures for Agricultural Research. 2nd ed. Wiley, New York, 680 p.

# Different degrees of freedom
mse <- c(6.73920, 1.93496, 1.15500, 10.58450)
df <- c(19, 16, 17, 19)

mse.bartlett.test(mse = c(6.73920, 1.93496, 1.15500, 10.58450),
                  df = c(19, 16, 17, 19))

# Same degrees of freedom
mse <- c(11.459848, 17.696970, 10.106818)
df <- c(20, 20, 20)

mse.bartlett.test(mse = c(11.459848, 17.696970, 10.106818),
                  df = c(20, 20, 20))
}
\seealso{
\code{\link[stats]{bartlett.test}}
}
