% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.augmentedRCBD.R
\name{pairwise.augmentedRCBD}
\alias{pairwise.augmentedRCBD}
\title{Perform Pairwise t Tests of Adjusted Means from \code{augmentedRCBD} Output}
\usage{
pairwise.augmentedRCBD(aug, cl = NULL, p.adjust = c("none", "tukey", "sidak"))
}
\arguments{
\item{aug}{An object of class \code{augmentedRCBD}.}

\item{cl}{A cluster object created by \code{\link[parallel]{makeCluster}} for
parallel evaluations.}

\item{p.adjust}{The p value adjustment method. Either \code{"none"},
\code{"tukey"} or \code{"sidak"}.}
}
\value{
A data frame of pairwise comparisons of treatments.
}
\description{
\code{pairwise.augmentedRCBD} performs pairwise t tests of adjusted means from
an object of class \code{augmentedRCBD}.
}
\details{
The default pairwise comparison in \code{\link[augmentedRCBD]{augmentedRCBD}}
employs \code{\link[emmeans]{pairs.emmGrid}} function from
\code{\link[emmeans]{emmeans}} which is very slow for large number of
comparisons. This function attempts to do the same faster with parallel
computing with the package \code{\link[parallel]{parallel-package}}.
}
\examples{
library(augmentedRCBD)
blk <- c(rep(1,7),rep(2,6),rep(3,7))
trt <- c(1, 2, 3, 4, 7, 11, 12, 1, 2, 3, 4, 5, 9, 1, 2, 3, 4, 8, 6, 10)
y1 <- c(92, 79, 87, 81, 96, 89, 82, 79, 81, 81, 91, 79, 78, 83, 77, 78, 78,
        70, 75, 74)
y2 <- c(258, 224, 238, 278, 347, 300, 289, 260, 220, 237, 227, 281, 311, 250,
        240, 268, 287, 226, 395, 450)
data <- data.frame(blk, trt, y1, y2)
# Convert block and treatment to factors
data$blk <- as.factor(data$blk)
data$trt <- as.factor(data$trt)
# Results for variable y1
out1 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE)
# Results for variable y2
out2 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE)

# Make cluster
ncores <- max(2, parallel::detectCores() - 2)
cl <- makeCluster(getOption("cl.cores", ncores))

# Pairwise t test without p value adjustment
pout1 <- pairwise.augmentedRCBD(out1, cl = NULL,
                                p.adjust = c("none"))
pout2 <- pairwise.augmentedRCBD(out1, cl = NULL,
                                p.adjust = c("none"))

# Pairwise t test with tukey adjustment
pout1_tukey <- pairwise.augmentedRCBD(out1, cl = NULL,
                                      p.adjust = "tukey")
pout2_tukey <- pairwise.augmentedRCBD(out1, cl = NULL,
                                      p.adjust = "tukey")

# Pairwise t test with sidak p value adjustment
pout1 <- pairwise.augmentedRCBD(out1, cl = NULL,
                                p.adjust = "sidak")
pout2 <- pairwise.augmentedRCBD(out1, cl = NULL,
                                p.adjust = "sidak")
}
\seealso{
\code{\link[parallel]{makeCluster}}
}
