% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mstrat_out.R
\name{parse_mstrat_out}
\alias{parse_mstrat_out}
\title{Parse Output Files from \code{MStrat}}
\usage{
parse_mstrat_out(
  data.file,
  genotype = NULL,
  variable.file,
  kernel.file,
  redundancy.output = NULL,
  core.output = NULL
)
}
\arguments{
\item{data.file}{The path to the \code{.dat} input file used for generating
the output files from \code{MStrat}.}

\item{genotype}{Name of column/variable with the genotype names as a
character string.}

\item{variable.file}{The path to the \code{.var} input file used for
generating the output files from \code{MStrat}.}

\item{kernel.file}{The path to the \code{.ker} input file used for generating
the output files from \code{MStrat}.}

\item{redundancy.output}{The path to the redundancy output file generated by
\code{MStrat}.}

\item{core.output}{The path to the core output file generated by
\code{MStrat}.}
}
\description{
Parse Output Files from \code{MStrat}
}
