% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_tests.R
\name{permutation_tests}
\alias{permutation_tests}
\alias{perm.test.global}
\alias{perm.test.pairwise}
\title{Permutation Tests}
\usage{
perm.test.global(
  x,
  group,
  fun,
  R = 1000,
  parallel = c("no", "multicore", "snow"),
  ncpus = 1L,
  cl = NULL,
  ...
)

perm.test.pairwise(
  x,
  group,
  fun,
  R = 1000,
  p.adjust.method = c("bonferroni", "holm"),
  parallel = c("no", "multicore", "snow"),
  ncpus = 1L,
  cl = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric or factor vector of observations.}

\item{group}{A factor vector indicating the group of each observation. Must
have the same length as \code{x}.}

\item{fun}{A function to summarize values within each group.}

\item{R}{Integer specifying the number of permutations. Default is 1000.}

\item{parallel}{
    The type of parallel operation to be used (if any).  If missing, the
    default is taken from the option \code{"boot.parallel"} (and if that
    is not set, \code{"no"}).
  }

\item{ncpus}{
    integer: number of processes to be used in parallel operation:
    typically one would chose this to the number of available CPUs.
  }

\item{cl}{
    An optional \pkg{parallel} or \pkg{snow} cluster for use if
    \code{parallel = "snow"}.  If not supplied, a cluster on the
    local machine is created for the duration of the \code{boot} call.
  }

\item{...}{Additional arguments passed to \code{fun}.}

\item{p.adjust.method}{(perm.test.pairwise only) Method for adjusting
p-values for multiple comparisons. Options include \code{"bonferroni"} and
\code{"holm"}. Default is \code{"bonferroni"}.}
}
\value{
\describe{ \item{\code{perm.test.global}}{A list of the following
  elements. \describe{ \item{test_stat}{The test statistic value.}
  \item{observed_values}{The observed values for each group.}
  \item{p_value}{The p value.} } } \item{\code{perm.test.pairwise}}{A data
  frame of the following columns. \describe{
  \item{Comparison}{The comparison.} \item{p.value}{The p value.}
  \item{adj.p.value}{The adjusted p value.} } } }
}
\description{
These functions perform permutation-based hypothesis tests to compare groups
with respect to a summary statistic (e.g., mean, diversity index).
\itemize{ \item{\code{perm.test.global} performs a global test across all
groups simultaneously, using a weighted sum of squares between group summary
indices as the test statistic} \item{\code{perm.test.pairwise} performs
pairwise tests between all combinations of groups, comparing the absolute
difference in summary statistics and optionally adjusting p-values for
multiple comparisons.} }
}
\examples{
library(EvaluateCore)

pdata <- cassava_CC

qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

# Conver '#t qualitative data columns to factor
pdata[, qual] <- lapply(pdata[, qual], as.factor)

str(pdata)

# Global tests ----

perm.test.global(x = pdata$NMSR, group = pdata$CUAL, fun = mean,
                 R = 100)

perm.test.global(x = pdata$LNGS, group = pdata$CUAL, fun = shannon,
                 R = 100)

perm.test.global(x = pdata$PTLC, group = pdata$CUAL, fun = simpson,
                 R = 100)

# Pairwise tests ----

perm.test.pairwise(x = pdata$NMSR, group = pdata$CUAL, fun = mean,
                   R = 100)

perm.test.pairwise(x = pdata$LNGS, group = pdata$CUAL, fun = shannon,
                   R = 100)

perm.test.pairwise(x = pdata$PTLC, group = pdata$CUAL, fun = simpson,
                   R = 100)

}
