% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_powercore_input.R
\name{prep_powercore_input}
\alias{prep_powercore_input}
\title{Prepare Input Files for \code{PowerCore}}
\usage{
prep_powercore_input(
  data,
  genotype,
  qualitative,
  quantitative,
  center = TRUE,
  scale = TRUE,
  always.selected = NULL,
  file.name = "PowerCore_input",
  folder.path = getwd()
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
columns with the genotype names and multiple quantitative and/or
qualitative trait/variable data.}

\item{genotype}{Name of column with the genotype names as a character string.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{center}{either a logical value or numeric-alike vector of length
    equal to the number of columns of \code{x}, where
    \sQuote{numeric-alike} means that \code{\link[base]{as.numeric}(.)} will
    be applied successfully if \code{\link[base]{is.numeric}(.)} is not true.}

\item{scale}{either a logical value or a numeric-alike vector of length
    equal to the number of columns of \code{x}.}

\item{always.selected}{A character vector with names of individuals in the
\code{genotype} that should always be selected in the core collection. The
maximum length accepted by \code{MStrat} is 500.}

\item{file.name}{A character string of name of file where the data will be
saved.}

\item{folder.path}{The path to folder where the input files are to be saved.}
}
\description{
Prepare input files for \code{PowerCore}, a program applying the advanced M
strategy with a heuristic search for establishing core sets
\insertCite{kim_PowerCore_2007,kim_PowerCore_2007a}{avial}.
}
\examples{

library(EvaluateCore)

data(cassava_EC)
data <- cassava_EC

quant <- c("NMSR", "TTRN", "TFWSR", "TTSW", "TTPW", "AVPW",
           "ARSR", "SRDM")
qual <- c("CUAL", "LNGS", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

# Prepare genotype column
data$Accession <- rownames(data)
rownames(data) <- NULL
data$Accession <- as.factor(data$Accession)

# Convert qualitative data as factors
data[, qual] <- lapply(data[, qual],
                       function(x) factor(as.factor(x)))

sel <- c("TMe-2906", "TMe-3412", "TMe-1374", "TMe-768", "TMe-14",
         "TMe-3284", "TMe-937", "TMe-1859", "TMe-3265", "TMe-1739",
         "TMe-972", "TMe-769", "TMe-3243", "TMe-3719", "TMe-1095",
         "TMe-893", "TMe-1262", "TMe-2083", "TMe-376", "TMe-3633",
         "TMe-1738", "TMe-2428", "TMe-259", "TMe-3457", "TMe-1406",
         "TMe-977", "TMe-3006", "TMe-925", "TMe-3671", "TMe-2779",
         "TMe-1293", "TMe-268", "TMe-266", "TMe-3562", "TMe-801")

prep_powercore_input(data = data, genotype = "Accession",
                     qualitative = qual, quantitative = quant,
                     center = TRUE, scale = TRUE,
                     always.selected = sel,
                     file.name = "PowerCore_input",
                     folder.path = tempdir())

}
\references{
\insertAllCited{}
}
