% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_quant}
\alias{summary_quant}
\alias{summary_qual}
\title{Summary Statistics of Qualitative and Quantitative trait data}
\usage{
summary_quant(data, group, trait, out.format = c("long", "wide"))

summary_qual(data, group, trait, out.format = c("long", "wide"))
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
columns specifying the trait and group.}

\item{group}{Name of column specifying the group as a character string.}

\item{trait}{Name of column specifying the trait as a character string. The
trait column should be of type \code{"numeric"} for quantitative traits and
of type \code{"factor"} for qualitative traits.}

\item{out.format}{The output format.}
}
\value{
A \code{tibble} data.frame with the summary statistics.
}
\description{
Summary Statistics of Qualitative and Quantitative trait data
}
