% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo-functions.R
\encoding{UTF-8}
\name{zenodo_gha_update}
\alias{zenodo_gha_update}
\title{Install a GitHub Actions workflow to create a \code{.zenodo.json} file}
\usage{
zenodo_gha_update(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Project root directory.}

\item{overwrite}{A logical value. If \code{TRUE}, overwrite an existing
workflow.}
}
\description{
This function installs a
  \href{https://github.com/features/actions}{GitHub Actions} workflow in your
  repository. The workflow updates your
  \href{https://help.zenodo.org/docs/github/describe-software/zenodo-json/}{\code{.zenodo.json}}
  file from the package \code{DESCRIPTION} when any of these events occur:
  \itemize{
    \item You publish a new release of the package.
    \item Your \code{DESCRIPTION} or \code{inst/CITATION} file is modified.
    \item The action can be run manually.
  }
}
\details{
Workflow triggers can be modified. See
  \href{https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows}{Events
  that trigger workflows}.
}
\examples{
\dontrun{
zenodo_gha_update()
}
}
\concept{git}
