% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.compare.R
\name{diversity.compare}
\alias{diversity.compare}
\title{Compare Diversity Measures}
\usage{
diversity.compare(
  x,
  group,
  R = 1000,
  base = exp(1),
  na.omit = TRUE,
  global.test = TRUE,
  pairwise.test = TRUE,
  bootstrap.ci = TRUE,
  diversity.profile = TRUE,
  p.adjust.method = c("bonferroni", "holm"),
  ci.conf = 0.95,
  ci.type = c("perc", "bca"),
  q = seq(0, 3, 0.1),
  parallel = c("no", "multicore", "snow"),
  ncpus = 1L,
  cl = NULL,
  seed = 123
)
}
\arguments{
\item{x}{A factor vector of categories (e.g., species, traits). The frequency
of each level is treated as the abundance of that category.}

\item{group}{A factor vector indicating the group of each observation. Must
have the same length as \code{x}.}

\item{R}{Integer specifying the number of permutations. Default is 1000.}

\item{base}{The logarithm base to be used for computation of shannon family
of diversity indices. Default is \code{exp(1)}.}

\item{na.omit}{logical. If \code{TRUE}, missing values (\code{NA}) are
ignored and not included as a distinct factor level for computation.
Default is \code{TRUE}.}

\item{global.test}{logical. If \code{TRUE} performs the global permutation
tests for the diversity measures. Default is \code{TRUE}.}

\item{pairwise.test}{logical. If \code{TRUE} performs the pairwise
permutation tests for the diversity measures. Default is \code{TRUE}.}

\item{bootstrap.ci}{logical. If \code{TRUE} computes the bootstrap confidence
intervals for the diversity measures. Default is \code{TRUE}.}

\item{diversity.profile}{logical. If \code{TRUE} diversity profiles. Default
is \code{TRUE}.}

\item{p.adjust.method}{(perm.test.pairwise only) Method for adjusting
p-values for multiple comparisons. Options include \code{"bonferroni"} and
\code{"holm"}. Default is \code{"bonferroni"}.}

\item{ci.conf}{Confidence level of the bootstrap interval. Default is 0.95.}

\item{ci.type}{A vector of character strings representing the type of
intervals required. The options are \code{c("perc", "bca")}.}

\item{q}{The order of the parametric index.}

\item{parallel}{
    The type of parallel operation to be used (if any).  If missing, the
    default is taken from the option \code{"boot.parallel"} (and if that
    is not set, \code{"no"}).
  }

\item{ncpus}{
    integer: number of processes to be used in parallel operation:
    typically one would chose this to the number of available CPUs.
  }

\item{cl}{
    An optional \pkg{parallel} or \pkg{snow} cluster for use if
    \code{parallel = "snow"}.  If not supplied, a cluster on the
    local machine is created for the duration of the \code{boot} call.
  }

\item{seed}{Integer. Random seed used to ensure reproducibility of
permutations and bootstrap. Default is 123.}
}
\value{
A list with the following elements. \describe{ \item{Diversity
  Indices}{A data frame of the different diversity indices computed for each
  group.}
  \item{Global Test}{A data frame of results of global permutation test
  including the test statistic (weighted sum of squares between group summary
indices) and the p value for the different diversity indices.}
  \item{Pairwise Test}{A list of the following data frames. \describe{
  \item{p-value}{A data frame of p values for each between
  group comparison for different diversity measures.}
  \item{cld}{A data frame of compact letter displays of significant
  differences among groups for different diversity measures.} } }
  \item{Bootstrap CIs}{A data frame of lower and upper bootstrap confidence
  intervals computed for each group in different diversity measures.}
  \item{Diversity profiles}{A list of data frames of Hill, Renyi and Tsallis
  diversity profiles computed for each group.} }
}
\description{
The function \code{diversity.compare} compares diversity indices between
different groups using the following approaches. \itemize{ \item{Global
permutation test across all groups simultaneously.} \item{Pairwise tests
between all combinations of groups.} \item{Bootstrap confidence intervals
(CI) for each group.} \item{Diversity profiles (Hill, Renyi and Tsallis) for
each group.}}
}
\note{
\itemize{ \item{In small samples with bounded statistics like Shannon
  Diversity Index and Menhinick index, the  bootstrap upper CI can equal the
  observed value because resamples cannot exceed the theoretical maximum.}
  \item{Similarly in small samples, the lower confidence bound can be zero
  because bootstrap resamples occasionally can contain only a single category
  (class or species), due to sampling uncertainty and the natural lower bound
  of the diversity index like Shannon Diversity Index.} \item{The BCa
  bootstrap can produce negative lower confidence limits due to boundary
  effects and skewness in the resampled distribution.} }
}
\examples{
library(EvaluateCore)

pdata <- cassava_CC

qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

# Convert qualitative data columns to factor
pdata[, qual] <- lapply(pdata[, qual], as.factor)

str(pdata)

diversity.compare(x = pdata$CUAL, group = pdata$LNGS, R = 100,
                  base = exp(1), na.omit = TRUE)

diversity.compare(x = pdata$ANGB, group = pdata$LNGS, R = 100,
                  base = exp(1), na.omit = TRUE)

}
