% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mstrat_out.R
\name{parse_mstrat_out}
\alias{parse_mstrat_out}
\title{Parse Output Files from \code{MStrat}}
\usage{
parse_mstrat_out(
  data.file,
  genotype = NULL,
  variable.file,
  kernel.file,
  redundancy.output = NULL,
  core.output = NULL
)
}
\arguments{
\item{data.file}{The path to the \code{.dat} input file used for generating
the output files from \code{MStrat}.}

\item{genotype}{Name of column/variable with the genotype names as a
character string.}

\item{variable.file}{The path to the \code{.var} input file used for
generating the output files from \code{MStrat}.}

\item{kernel.file}{The path to the \code{.ker} input file used for generating
the output files from \code{MStrat}.}

\item{redundancy.output}{The path to the redundancy output file generated by
\code{MStrat}.}

\item{core.output}{The path to the core output file generated by
\code{MStrat}.}
}
\value{
A list with the following components: \item{\code{MStrat Core
  Output}}{The raw output of core sets constructed by \code{MStrat}.}
  \item{\code{MStrat Core Optimised}}{The optimised core set output from
  \code{MStrat}.} \item{\code{MStrat Redundancy Output}}{The raw output of
  redundancy estimation by \code{MStrat}} \item{\code{MStrat Redundancy
  Plot}}{A list of plots of redundancy estimation.}
}
\description{
Prepare raw output files generated by \code{MStrat}
\insertCite{schoen_Conservation_1993a,gouesnard_MSTRAT_2001,gouesnard_MStrat_2002}{avial}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

parse_mstrat_out(data.file = "MStrat_input_data.dat",
                 genotype = "Accession",
                 variable.file = "MStrat_input_variable.var",
                 kernel.file = "MStrat_input_kernel.ker",
                 redundancy.output = "MStrat - Redundancy.out",
                 core.output = "MStrat - Core.out")
\dontshow{\}) # examplesIf}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[avial]{prep_mstrat_input}}
}
