% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_quant}
\alias{summary_quant}
\alias{summary_qual}
\title{Summary Statistics of Qualitative and Quantitative trait data}
\usage{
summary_quant(data, group = NULL, trait, out.format = c("long", "wide"))

summary_qual(data, group = NULL, trait, out.format = c("long", "wide"))
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
columns specifying the trait and group.}

\item{group}{Name of column specifying the group as a character string.}

\item{trait}{Name of column specifying the trait as a character string. The
trait column should be of type \code{"numeric"} for quantitative traits and
of type \code{"factor"} for qualitative traits.}

\item{out.format}{The output format.}
}
\value{
A \code{tibble} data.frame with the summary statistics.
}
\description{
Summary Statistics of Qualitative and Quantitative trait data
}
\examples{
library(agridat)
library(dplyr)

soydata <- australia.soybean

soydata$year <- as.factor(soydata$year)

quant_traits <- c("yield", "height", "lodging",
                  "size", "protein", "oil")
set.seed(123)
soydata <-
  soydata \%>\%
  mutate(
    across(
      .cols = all_of(quant_traits),
      .fns = ~factor(cut(.x, breaks = quantile(.x, na.rm = TRUE),
                         include.lowest = TRUE),
                     labels = sample(1:9, size = 4)),
      .names = "{.col}_score"
    )
  )

qual_traits <- c("yield_score", "height_score", "lodging_score",
                 "size_score", "protein_score", "oil_score")

summary_quant(data = soydata, trait = quant_traits)
summary_qual(data = soydata, trait = qual_traits)

summary_quant(data = soydata, group = "loc", trait = quant_traits)
summary_qual(data = soydata, group = "loc", trait = qual_traits)

summary_quant(data = soydata, group = c("loc", "year"), trait = quant_traits)
summary_qual(data = soydata, group = c("loc", "year"), trait = qual_traits)

}
