% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_barwidth_grid.R
\name{uniform_barwidth_grid}
\alias{uniform_barwidth_grid}
\title{Plot a List of \code{ggplot2} Barplots as a Grid with with Uniform Width of
the Bars}
\usage{
uniform_barwidth_grid(plot.list, level.count, nrow, ncol)
}
\arguments{
\item{plot.list}{The list of bar plots as \code{ggplot2} objects.}

\item{level.count}{The count of levels/bars in each plot.}

\item{nrow}{Number of rows of the grid.}

\item{ncol}{Number of columns of the grid.}
}
\value{
The plot grid as a \code{patchwork} object.
}
\description{
Plot a List of \code{ggplot2} Barplots as a Grid with with Uniform Width of
the Bars
}
